Move the interfaces into their own files. This will be useful when we share them between the .NET Core and .NET Framework WebAPI
