+++ src/Jackett/Jackett.csproj
@@ -203,6 +203,21 @@
     <Compile Include="Models\TrackerCache.cs" />
     <Compile Include="Models\TrackerCacheResult.cs" />
     <Compile Include="Services\CacheService.cs" />
+    <Compile Include="Services\Interfaces\ICacheService.cs" />
+    <Compile Include="Services\Interfaces\IImdbResolver.cs" />
+    <Compile Include="Services\Interfaces\IConfigurationService.cs" />
+    <Compile Include="Services\Interfaces\IIndexerConfigurationService.cs" />
+    <Compile Include="Services\Interfaces\IIndexerManagerService.cs" />
+    <Compile Include="Services\Interfaces\ILogCacheService.cs" />
+    <Compile Include="Services\Interfaces\IProcessService.cs" />
+    <Compile Include="Services\Interfaces\IProtectionService.cs" />
+    <Compile Include="Services\Interfaces\ISecuityService.cs" />
+    <Compile Include="Services\Interfaces\ISerializeService.cs" />
+    <Compile Include="Services\Interfaces\IServerService.cs" />
+    <Compile Include="Services\Interfaces\IRunTimeService.cs" />
+    <Compile Include="Services\Interfaces\IServiceConfigService.cs" />
+    <Compile Include="Services\Interfaces\ITrayLockService.cs" />
+    <Compile Include="Services\Interfaces\IUpdateService.cs" />
     <Compile Include="Services\LogCacheService.cs" />
     <Compile Include="Services\ProtectionService.cs" />
     <Compile Include="Services\TrayLockService.cs" />
