+++ src/Jackett/Services/ConfigurationService.cs
@@ -15,20 +15,6 @@ using System.Windows.Forms;
 
 namespace Jackett.Services
 {
-    public interface IConfigurationService
-    {
-        string GetContentFolder();
-        string GetVersion();
-        string GetIndexerConfigDir();
-        string GetAppDataFolder();
-        string GetSonarrConfigFile();
-        T GetConfig<T>();
-        void SaveConfig<T>(T config);
-        string ApplicationFolder();
-        List<string> GetCardigannDefinitionsFolders();
-        void CreateOrMigrateSettings();
-        void PerformMigration();
-    }
 
     public class ConfigurationService : IConfigurationService
     {
