+++ src/Jackett/Services/Interfaces/IConfigurationService.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+
+namespace Jackett.Services.Interfaces
+{
+    public interface IConfigurationService
+    {
+        string GetContentFolder();
+        string GetVersion();
+        string GetIndexerConfigDir();
+        string GetAppDataFolder();
+        string GetSonarrConfigFile();
+        T GetConfig<T>();
+        void SaveConfig<T>(T config);
+        string ApplicationFolder();
+        List<string> GetCardigannDefinitionsFolders();
+        void CreateOrMigrateSettings();
+        void PerformMigration();
+    }
+}
