+++ src/Jackett/Services/Interfaces/IImdbResolver.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using Jackett.Utils;
+
+namespace Jackett.Services.Interfaces
+{
+    public interface IImdbResolver
+    {
+        Task<Movie> MovieForId(NonNull<string> imdbId);
+    }
+}
