+++ src/Jackett/Services/Interfaces/ILogCacheService.cs
@@ -1,0 +1,11 @@
+﻿using Jackett.Models;
+using System.Collections.Generic;
+
+namespace Jackett.Services.Interfaces
+{
+    public interface ILogCacheService
+    {
+       // void AddLog(LogEventInfo l);
+        List<CachedLog> Logs { get; }
+    }
+}
