+++ src/Jackett/Services/Interfaces/IProtectionService.cs
@@ -1,0 +1,10 @@
+﻿namespace Jackett.Services.Interfaces
+{
+    public interface IProtectionService
+    {
+        byte[] InstanceKey { get; set; }
+
+        string Protect(string plainText);
+        string UnProtect(string plainText);
+    }
+}
