+++ src/Jackett/Services/Interfaces/IRunTimeService.cs
@@ -1,0 +1,7 @@
+﻿namespace Jackett.Services.Interfaces
+{
+    public interface IRunTimeService
+    {
+        void Spin();
+    }
+}
