+++ src/Jackett/Services/Interfaces/ISerializeService.cs
@@ -1,0 +1,8 @@
+﻿namespace Jackett.Services.Interfaces
+{
+    public interface ISerializeService
+    {
+        string Serialise(object obj);
+        T DeSerialise<T>(string json);
+    }
+}
