+++ src/Jackett/Services/Interfaces/IServiceConfigService.cs
@@ -1,0 +1,12 @@
+﻿namespace Jackett.Services.Interfaces
+{
+    public interface IServiceConfigService
+    {
+        void Install();
+        void Uninstall();
+        bool ServiceExists();
+        bool ServiceRunning();
+        void Start();
+        void Stop();
+    }
+}
