+++ src/Jackett/Services/Interfaces/ITrayLockService.cs
@@ -1,0 +1,8 @@
+﻿namespace Jackett.Services.Interfaces
+{
+    public interface ITrayLockService
+    {
+        void WaitForSignal();
+        void Signal();
+    }
+}
