+++ src/Jackett/Services/Interfaces/IUpdateService.cs
@@ -1,0 +1,9 @@
+﻿namespace Jackett.Services.Interfaces
+{
+    public interface IUpdateService
+    {
+        void StartUpdateChecker();
+        void CheckForUpdatesNow();
+        void CleanupTempDir();
+    }
+}
