+++ src/Jackett/Services/ProcessService.cs
@@ -8,11 +8,6 @@ using System.Threading.Tasks;
 
 namespace Jackett.Services
 {
-    public interface IProcessService
-    {
-        void StartProcessAndLog(string exe, string args, bool asAdmin = false);
-        string StartProcessAndGetOutput(string exe, string args, bool keepnewlines = false, bool asAdmin = false);
-    }
 
     public class ProcessService : IProcessService
     {
