+++ src/Jackett/Services/ServerService.cs
@@ -20,19 +20,6 @@ using System.Web;
 
 namespace Jackett.Services
 {
-    public interface IServerService
-    {
-        void Initalize();
-        void Start();
-        void Stop();
-        void ReserveUrls(bool doInstall = true);
-        ServerConfig Config { get; }
-        void SaveConfig();
-        Uri ConvertToProxyLink(Uri link, string serverUrl, string indexerId, string action = "dl", string file = "t.torrent");
-        string BasePath();
-        string GetServerUrl(HttpRequestMessage Request);
-        List<string> notices { get; } 
-    }
 
     public class ServerService : IServerService
     {
